requirements.txt
# Hand Tracking Media Controller Requirements
# Install with: pip install -r requirements.txt

# Computer Vision & Hand Tracking
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3

# Web Automation
selenium==4.15.2
webdriver-manager==4.0.1

# Utility Libraries
python-dotenv==1.0.0
psutil==5.9.6
requests==2.31.0

# Logging and Configuration
pathlib2==2.3.7

# Optional - for better performance
# Uncomment if you want to use these
scikit-learn==1.3.0
pandas==2.0.3